+++ job_data/transformation.py
@@ -1,22 +1,24 @@
 import re
 import pandas as pd
 from pandas import Series
-from typing import List
 
 DATE_FORMAT = '%d/%m/%Y'
 SEPARATE_TECHNOLOGIES = ' - '
 REGEX_EXTRACT_AMOUNT = r'(\d+(\.\d{3})*\€(?:\s*-\s*\d+(\.\d{3})*\€)?)|(\d+(\.\d{3})*\€)'
 REGEX_MODALIDAD = r'(presencial|hibrido|híbrido|remoto)'
 
+
 def add_date(value: Series) -> Series:
     value = value.strip()
     value = pd.to_datetime(value[:10], format=DATE_FORMAT)
     return value
 
-def add_list_technologies(value: str) -> List[str]:
+
+def add_list_technologies(value: str) -> list[str]:
     return value.split(SEPARATE_TECHNOLOGIES)
 
-def extract_salary_amount(descripcion: str) -> str:
+
+def extract_salary_amount(descripcion: str) -> str | None:
     matches = re.findall(REGEX_EXTRACT_AMOUNT, descripcion)
     if matches:
         if matches[0][0]:
@@ -25,30 +27,39 @@ def extract_salary_amount(descripcion: str) -> str:
             return matches[0][3]
     return None
 
-def extract_salary_min(salary: str) -> float:
+
+def extract_salary_min(salary: str) -> float | None:
     if salary is None:
         return None
-
     if '-' in salary:
         return float(salary.split(' - ')[0].replace('€', '').replace('.', ''))
     else:
         return float(salary.replace('€', '').replace('.', ''))
 
-def extract_salary_max(salary: str) -> float:
+
+def extract_salary_max(salary: str) -> float | None:
     if salary is None:
         return None
-
     if '-' in salary:
         return float(salary.split(' - ')[1].replace('€', '').replace('.', ''))
     else:
         return float(salary.replace('€', '').replace('.', ''))
 
-def extract_modality(descripcion: str) -> str:
-    matches = re.findall(REGEX_MODALIDAD, descripcion)
+
+def extract_modality(description: str) -> str | None:
+    matches = re.findall(REGEX_MODALIDAD, description)
     if matches:
         return matches[0].capitalize()
     else:
         return None
 
-def normalized_technologies(technologies: list) -> list:
-    return []
+
+# def normalized_technology_name(raw_name: str, normal: dict) -> str:
+#     if raw_name in normal.keys():
+#         return normal[raw_name]
+#     else:
+#         return raw_name
+#
+#
+# def normalized_technologies(technologies: list, normalized_names: dict) -> list:
+#     return [normalized_technology_name(technology, normalized_names) for technology in technologies]