+++ CHANGELOG.md
@@ -1,2 +1,15 @@
+
+## Version 0.1.0
+Features:
+- Add job_data package
+- Add tests
+- Add data executions folder
+- Add config folder
+- Add .gitignore
+- Add README.md
+- Add requirements.txt
+- Add setup.py
+- Add main.py
+
 ## Version 0.0.1
 Create initial project version